Show popover views in overlaid UIView atop window
